refactor(data.tests): Added api specific tests
I split data.api into data.api and data.etl.  Etl handles the actual
data manipulation while data.api handles the requests and logic
associated with the open data portal.  This is preliminary work for
handling temporal data collection.

Issue #87
